+++ bin/rmemo.js
@@ -19,6 +19,7 @@ import { cmdWatch } from "../src/cmd/watch.js";
 import { cmdWs } from "../src/cmd/ws.js";
 import { cmdProfile } from "../src/cmd/profile.js";
 import { cmdSession } from "../src/cmd/session.js";
+import { cmdServe } from "../src/cmd/serve.js";
 import { parseArgs, printHelp } from "../src/lib/args.js";
 import { exitWithError } from "../src/lib/io.js";
 
@@ -86,6 +87,9 @@ try {
       break;
     case "session":
       await cmdSession({ rest, flags });
+      break;
+    case "serve":
+      await cmdServe({ flags });
       break;
     case "help":
     case undefined: